# openapi-snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HLOOL Mail API",
    "version": "0.8.0",
    "description": "由内部 apispec 注册表生成的机器可读 API 元数据。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "已配置的公开基础 URL"
    }
  ],
...
  "paths": {
    "/api/docs.md": {
      "get": {
        "tags": [
          "公开元数据"
        ],
        "summary": "Markdown API 指南",
        "description": "返回由 API 注册表生成、适合 AI 阅读的 Markdown 指南。",
        "operationId": "apiDocsMarkdown",
        "responses": {
          "200": {
            "description": "Markdown 文档",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "public"
      }
    },
    "/api/domains/available": {
      "get": {
        "tags": [
          "API Key 自动化"
        ],
        "summary": "可用域名",
        "description": "列出可选择的公共域名以及 API Key 拥有者可访问的私有域名。客户端和 AI 助手应把该端点作为生成邮箱前的域名选择入口，优先使用 data.public_domains 与 data.private_domains；data.domains 仅为旧客户端公共域名兼容字段。",
        "operationId": "availableDomains",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeAvailableDomains"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "X-API-Key"
      }
    },
...
    "/api/openapi.json": {
      "get": {
        "tags": [
          "公开元数据"
        ],
        "summary": "OpenAPI JSON",
        "description": "返回 JSON 格式的 OpenAPI 文档。",
        "operationId": "openAPIJSON",
        "responses": {
          "200": {
            "description": "OpenAPI JSON 文档",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "public"
      }
    },
    "/api/openapi.yaml": {
      "get": {
        "tags": [
          "公开元数据"
        ],
        "summary": "OpenAPI YAML",
        "description": "返回 YAML 格式的 OpenAPI 文档。",
        "operationId": "openAPIYAML",
        "responses": {
          "200": {
            "description": "OpenAPI YAML 文档",
            "content": {
              "application/yaml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "public"
      }
    },
...
    "/api/skill.md": {
      "get": {
        "tags": [
          "公开元数据"
        ],
        "summary": "Skill 指南",
        "description": "返回由 API 注册表生成的 AI 助手 Skill 指南。",
        "operationId": "apiSkillMarkdown",
        "responses": {
          "200": {
            "description": "Markdown Skill 文档",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "public"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailableDomain": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "domain": {
            "type": "string"
          },
          "id": {
            "format": "uint64",
            "type": "integer"
          },
          "message_count": {
            "format": "int64",
            "type": "integer"
          },
          "mode": {
            "type": "string"
          },
          "root_ready": {
            "type": "boolean"
          },
          "wildcard_pattern": {
            "type": "string"
          },
          "wildcard_ready": {
            "type": "boolean"
          },
          "wildcard_requested": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "domain",
          "mode",
          "message_count"
        ],
        "type": "object"
      },
      "AvailableDomainsData": {
        "additionalProperties": false,
        "properties": {
          "domains": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "private_domains": {
            "items": {
              "$ref": "#/components/schemas/AvailableDomain"
            },
            "type": "array"
          },
          "public_domains": {
            "items": {
              "$ref": "#/components/schemas/AvailableDomain"
            },
            "type": "array"
          },
          "public_unavailable_reason": {
            "type": "string"
          }
        },
        "required": [
          "domains",
          "public_domains",
          "private_domains"
        ],
        "type": "object"
      },
...
      "EnvelopeAvailableDomains": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AvailableDomainsData"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "nullable": true
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
...
      "Usage": {
        "additionalProperties": false,
        "properties": {
          "daily_limit": {
            "type": "string"
          },
          "daily_unlimited": {
            "type": "string"
          },
          "remaining_today": {
            "type": "string"
          },
          "remaining_total": {
            "type": "string"
          },
          "total_limit": {
            "type": "string"
          },
          "total_unlimited": {
            "type": "string"
          },
          "total_usage": {
            "type": "string"
          },
          "used_today": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "自动化端点使用的 API Key。"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gptmail_session",
        "description": "Web Console 会话专用管理端点使用的会话 Cookie。"
      }
    }
  },
...
}